requirements.txt
streamlit
volcengine-python-sdk[ark]
python-dotenv 